requirements.txt
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.10.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wcwidth==0.2.6
webdriver-manager==4.0.1
wsproto==1.2.0
zipp==3.17.0
